finetune/backend/requirements.txt
fastapi
uvicorn
torch
transformers
unsloth
bitsandbytes
accelerate
pydantic
python-dotenv
paramiko
scp 
openpyxl